webhook_listener/requirements.txt
discord~=1.7.3
loguru~=0.5.3
requests
python-dotenv~=0.19.2
gunicorn~=20.1.0
flask~=2.0.2
numpy